+++ pruebas/01-reading-list/byronjvh/README.md
@@ -1,2 +1,9 @@
+# Prueba Técnica Reading List
+Prueba técnica realizada: 
+[Reading list](https://github.com/midudev/pruebas-tecnicas/tree/main/pruebas/01-reading-list)
+
+Prueba técnica de:
+[Pruebas técnicas de programación](https://pruebastecnicas.com/)
+
 ## Deploy
-* [GitHub pages](https://byronjvh.github.io/4Imagenes1Palabra/)
+* [Vercel](https://pruebas-tecnicas-murex.vercel.app/)
